Added recipient email variable
The variable "email_address" was added to remove the hardcoded email address used for a previous implementation of this script. The new variable represents the main recipient of any emails, this is essentially your own email, however you might receive emails sent to another address, in which case this variable should be the email of interest to you.